Add the Text Styles section css
This section was less of a challenge compared to the Fonts and
Colors sections. I was a bit more comfortable using Flexbox,
and although I know I still made a lot of mistakes I'm sure
I'll be able to identify them in the near future.

This is also the first time I use git commit to open the text
editor and make a commit with a message.